[Fix] InquirtAlarmEntity 오타 수정, 신고 알람은 삭제 안되도록 수정, ReportFilter -> re…

+++ src/main/java/hs/kr/backend/devpals/domain/user/service/UserAlarmService.java
@@ -99,9 +99,10 @@ public class UserAlarmService {
         AlarmEntity alarmEntity
                 = alarmRepository.findByReceiverIdAndAlarmId(userId, alarmId).orElseThrow(() -> new CustomException(ErrorException.ALARM_NOT_FOUND));
 
-        //TODO: 신고는 삭제 안되게 해야함
-//        if(alarmEntity.getAlarmFilterIntValue().equals(AlarmFilter.APPLIED_PROJECTS.getValue());
-//            throw new CustomException(ErrorException.CAN_NOT_DELETE_ALARM);
+        //신고알람은 삭제 안되게 설정
+        if(alarmEntity.getAlarmFilterIntValue().equals(AlarmFilter.REPORT.getValue())) {
+            throw new CustomException(ErrorException.CAN_NOT_DELETE_ALARM);
+        }
         alarmRepository.delete(alarmEntity);
 
         refreshCacheUserAlarm(userId);
